+++ include/UECS/detail/Chunk.h
@@ -2,7 +2,6 @@
 
 #include <cstdint>
 #include <array>
-#include <tuple>
 #include <vector>
 
 #include "../config.h"