+++ include/UECS/detail/CmptTypeSet.h
@@ -1,10 +1,8 @@
 #pragma	once
 
-#include "Util.h"
 #include "../EntityQuery.h"
 
-#include <UTemplate/TypeID.h>
-#include <UTemplate/TemplateList.h>
+#include <UContainer/Span.h>
 
 #include <set>
 
@@ -14,15 +12,15 @@ namespace Ubpa::UECS {
 
 		size_t HashCode() const noexcept;
 
-		void Insert(const CmptType* types, size_t num);
-		void Erase(const CmptType* types, size_t num) noexcept;
+		void Insert(Span<const CmptType> types);
+		void Erase(Span<const CmptType> types) noexcept;
 		bool Contains(CmptType type) const;
 
-		bool ContainsAll(const CmptType* types, size_t num) const;
+		bool ContainsAll(Span<const CmptType> types) const;
 		template<typename CmptTypeContainer>
 		bool ContainsAll(const CmptTypeContainer& types) const;
 
-		bool ContainsAny(const CmptType* types, size_t num) const;
+		bool ContainsAny(Span<const CmptType> types) const;
 		template<typename CmptTypeContainer>
 		bool ContainsAny(const CmptTypeContainer& types) const;