+++ doc/changelog.md
@@ -1,6 +1,8 @@
 # Change Log
 
-- 0.14.3: `CmptsView`, `SingletonsView` use `Span` 
+- 0.14.3
+  - `CmptsView`, `SingletonsView` use `Span` 
+  - API with `const CmptType* types, size_t num` use `Span<CmptType> types` as instead
 - 0.14.2
   - add `RandomAccessor` for random access other entity's component
     - `SystemFunc` add the member